wooters

+++ src/aolib.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void delay(volatile uint32_t delay);
+void delay(volatile uint64_t delay);
 
 #endif
